+++ cmd/shortener/main.go
@@ -12,9 +12,6 @@ func init() {
 }
 
 func main() { //nolint:typecheck
-	//log.Println("[INFO] main -> cfg.GetFlag()")
-	//cfg.GetFlag()
-
 	log.Println("[INFO] main -> cfg.GetEnvs()")
 	if err := cfg.GetEnvs(); err != nil {
 		log.Fatal(err)
