+++ internal/app/server/handlers/get_all_pairs_for_user_test.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"github.com/EestiChameleon/URLShortenerService/internal/app/cfg"
-	resp "github.com/EestiChameleon/URLShortenerService/internal/app/responses"
+	resp "github.com/EestiChameleon/URLShortenerService/internal/app/server/responses"
 	"github.com/EestiChameleon/URLShortenerService/internal/app/storage"
 	"github.com/go-chi/chi/v5"
 	"github.com/stretchr/testify/assert"
@@ -15,44 +15,35 @@ import (
 	"testing"
 )
 
-func TestGetOrigURL(t *testing.T) {
+func TestGetAllPairs(t *testing.T) {
 	type want struct {
-		contentType    string
-		statusCode     int
-		respMessage    string
-		headerLocation string
+		contentType string
+		statusCode  int
+		respMessage string
 	}
 
 	tests := []struct {
 		name    string
 		request string
+		userID  string
 		want    want
 	}{
 		{
-			name:    "GET test #1: test url -> 307",
-			request: "/test",
+			name:    "GET test #1: no user id - no content -> 204",
+			request: "http://localhost:8080/api/user/urls",
+			userID:  "",
 			want: want{
-				contentType:    resp.MIMETextPlainCharsetUTF8,
-				statusCode:     307,
-				headerLocation: "https://jwt.io/",
+				statusCode: 204,
 			},
 		},
 		{
-			name:    "GET test #2: empty id -> 404",
-			request: "/",
+			name:    "GET test #2: test user id -> 200",
+			request: "http://localhost:8080/api/user/urls",
+			userID:  "testUser",
 			want: want{
-				contentType: resp.MIMETextPlainCharsetUTF8,
-				statusCode:  404,
-				respMessage: "404 page not found\n",
-			},
-		},
-		{
-			name:    "GET test #3: wrong id -> 400",
-			request: "/666xxx",
-			want: want{
-				contentType: resp.MIMETextPlainCharsetUTF8,
-				statusCode:  400,
-				respMessage: "invalid id",
+				contentType: resp.MIMEApplicationJSONCharsetUTF8,
+				statusCode:  200,
+				respMessage: "[{\"short_url\":\"http://localhost:8080/test\",\"original_url\":\"https://jwt.io/\"}]",
 			},
 		},
 	}
@@ -66,20 +57,22 @@ func TestGetOrigURL(t *testing.T) {
 			if err := storage.InitStorage(); err != nil {
 				log.Fatal(err)
 			}
-			if tt.request == "/test" {
+			if tt.userID != "" {
 				storage.User.SetUserID("testUser")
-				storage.User.SavePair(storage.Pair{
+				if err := storage.User.SavePair(storage.Pair{
 					ShortURL: "http://localhost:8080/test",
 					OrigURL:  "https://jwt.io/",
-				})
+				}); err != nil {
+					log.Fatal(err)
+				}
 			}
 			defer os.Remove(cfg.Envs.FileStoragePath)
-
 			// создаём новый Recorder
 			w := httptest.NewRecorder()
 			r := chi.NewRouter()
 			// определяем хендлер
-			r.Get("/{id}", GetOrigURL)
+			storage.User.SetUserID(tt.userID)
+			r.Get("/api/user/urls", GetAllPairs)
 			// запускаем сервер
 			r.ServeHTTP(w, request)
 			res := w.Result()
