+++ internal/app/server/handlers/post_batch.go
@@ -3,7 +3,7 @@ package handlers
 import (
 	"encoding/json"
 	"errors"
-	resp "github.com/EestiChameleon/URLShortenerService/internal/app/responses"
+	resp "github.com/EestiChameleon/URLShortenerService/internal/app/server/responses"
 	"github.com/EestiChameleon/URLShortenerService/internal/app/service/process"
 	"github.com/EestiChameleon/URLShortenerService/internal/app/storage"
 	"io"
