+++ internal/app/server/handlers/post_provide_short_url.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"errors"
-	resp "github.com/EestiChameleon/URLShortenerService/internal/app/responses"
+	resp "github.com/EestiChameleon/URLShortenerService/internal/app/server/responses"
 	"github.com/EestiChameleon/URLShortenerService/internal/app/service/process"
 	"github.com/EestiChameleon/URLShortenerService/internal/app/storage"
 	"io"
