+++ internal/app/server/handlers/post_provide_short_url_test.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"github.com/EestiChameleon/URLShortenerService/internal/app/cfg"
-	resp "github.com/EestiChameleon/URLShortenerService/internal/app/responses"
+	resp "github.com/EestiChameleon/URLShortenerService/internal/app/server/responses"
 	"github.com/EestiChameleon/URLShortenerService/internal/app/storage"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -15,7 +15,7 @@ import (
 	"testing"
 )
 
-func TestJSONShortURL(t *testing.T) {
+func TestPostProvideShortURL(t *testing.T) {
 	type want struct {
 		contentType string
 		statusCode  int
@@ -30,24 +30,24 @@ func TestJSONShortURL(t *testing.T) {
 		want        want
 	}{
 		{
-			name:        "POST test #1: correct json -> 201",
-			contentType: resp.MIMEApplicationJSONCharsetUTF8,
-			request:     "http://localhost:8080/api/shorten",
-			body:        `{"url":"https://jwt.io/"}`,
+			name:        "POST test #1: url -> 201",
+			contentType: resp.MIMETextPlainCharsetUTF8,
+			request:     "http://localhost:8080/",
+			body:        "https://jwt.io/",
 			want: want{
-				contentType: resp.MIMEApplicationJSONCharsetUTF8,
+				contentType: resp.MIMETextPlainCharsetUTF8,
 				statusCode:  201,
 				respMessage: "",
 			},
 		},
 		{
-			name:        "POST test #2: wrong incoming json -> 400",
-			contentType: resp.MIMEApplicationJSONCharsetUTF8,
-			request:     "http://localhost:8080/api/shorten",
-			body:        `{"urly":"https://jwt.io/"}`,
+			name:        "POST test #2: empty url -> 400",
+			contentType: resp.MIMETextPlainCharsetUTF8,
+			request:     "http://localhost:8080/",
+			body:        "",
 			want: want{
 				contentType: resp.MIMETextPlainCharsetUTF8,
-				respMessage: "invalid data",
+				respMessage: "invalid url",
 				statusCode:  400,
 			},
 		},
@@ -59,12 +59,12 @@ func TestJSONShortURL(t *testing.T) {
 			// создаём новый Recorder
 			w := httptest.NewRecorder()
 			// определяем хендлер
-			h := http.HandlerFunc(JSONShortURL)
+			h := http.HandlerFunc(PostProvideShortURL)
+			// запускаем сервер
 			// envs
 			cfg.Envs.BaseURL = "http://localhost:8080"
 			//cfg.Envs.FileStoragePath = "tmp/testFile"
 			//cfg.Envs.DatabaseDSN = "postgresql://localhost:5432/yandex_practicum_db"
-			// запускаем сервер
 			if err := storage.InitStorage(); err != nil {
 				log.Fatal(err)
 			}
