+++ internal/app/server/responses/templates.go
